+++ depends/lua.mk
@@ -1,6 +1,3 @@
-# for mingw build on windows, use:
-# make -f lua.mk PLATFORM=generic INSTALLROOT=/d/workspace/github/bitmail/out
-
 package=lua
 package_version=5.3.4
 package_download_url=http://www.lua.org/ftp/lua-5.3.4.tar.gz
@@ -12,13 +9,18 @@ package_build_dir=$(package_extract_dir)/$(package)-$(package_version)
 package_output_dir=$${INSTALLROOT}
 package_platform=$${LUA_PLATFORM}
 
+# for mingw build on windows, use:
+# make -f lua.mk PLATFORM=generic INSTALLROOT=/d/workspace/github/bitmail/out
 
-all: $(package_tarball)
+
+all: download
 	tar zxvf $(package_tarball) -C $(package_extract_dir) && \
 	cd $(package_build_dir) && \
 	make $(package_platform) && \
 	make install INSTALL_TOP=$(package_output_dir) && \
 	make clean
 
-$(package_tarball):
-	curl -k $(package_download_url) -o $(package_tarball)
+download:
+	if ! [ -f $(package_tarball) ]; then curl -k $(package_download_url) -o $(package_tarball); fi;
+
+.PHONY : all download
